+++ test_unstructured/partition/test_code.py
@@ -4,6 +4,7 @@ import io
 import pytest
 
 from unstructured.partition.code import partition_code
+from unstructured.partition.auto import partition
 from unstructured.utils import get_homedir
 
 
@@ -29,6 +30,15 @@ TEST_FILE_PATH = {
     # "swift":
 }
 
+LANGUAGE_TO_EXT = {
+    "python": ".py",
+    "javascript": ".js",
+    "typescript": ".ts",
+    "cpp": ".cpp",
+    "c": ".c",
+    "go": ".go",
+}
+
 
 @pytest.mark.parametrize("language", TEST_FILE_PATH.keys())
 @pytest.fixture(scope="function")
@@ -49,11 +59,24 @@ def test_partition_code_from_filename(language: str, tmp_path_factory):
 @pytest.mark.parametrize("language", TEST_FILE_PATH.keys())
 def test_partition_code_from_file(language: str):
     file = io.BytesIO(open(TEST_FILE_PATH[language], "rb").read())
-    elements = partition_code(file=file)
+    elements = partition_code(file=file, programming_language=language)
     assert len(elements) > 0
     for el in elements:
         assert el.metadata.languages == [language]
     pass
+
+
+def test_programming_language_enforced():
+    with pytest.raises(ValueError):
+        partition_code(filename=TEST_FILE_PATH["python"], programming_language="go")
+
+
+@pytest.mark.parametrize("language", TEST_FILE_PATH.keys())
+def test_partition_auto_for_code(language: str):
+    file = io.BytesIO(open(TEST_FILE_PATH[language], "rb").read())
+    partition(
+        file=file, programming_language=language, file_filename="test" + LANGUAGE_TO_EXT[language]
+    )
 
 
 @pytest.mark.parametrize("max_partition", [500, 1000, 2000])
@@ -62,7 +85,8 @@ def test_partition_code_max_partition(max_partition: int, language: str):
     elements = partition_code(filename=TEST_FILE_PATH[language], max_partition=max_partition)
     assert len(elements) > 0
     for el in elements:
-        assert len(el.text) <= max_partition + 50
+        # The minimum is enforced, so there is a maximum spillage of 200
+        assert len(el.text) <= max_partition + 200
 
 
 @pytest.mark.parametrize("min_partition", [50, 100, 200])